+++ src/tools/dbcparser/SQLDMLGenerator.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Ember
+ * Copyright (c) 2019 - 2021 Ember
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -13,6 +13,7 @@
 #include <spark/buffers/BinaryStream.h>
 #include <spark/buffers/DynamicBuffer.h>
 #include <spark/buffers/VectorBufferAdaptor.h>
+#include <gsl/gsl_util>
 #include <fstream>
 #include <vector>
 #include <sstream>
@@ -245,7 +246,7 @@ std::vector<std::byte> load_dbc(const std::string& filename) {
 		throw std::runtime_error("Unable to open DBC for reading, " + filename);
 	}
 
-	std::vector<std::byte> data(size);
+	std::vector<std::byte> data(gsl::narrow<std::size_t>(size));
 	file.read(reinterpret_cast<char*>(data.data()), data.size());
 	return data;
 }
